Extracted Create Offer code into Editable Offer abstract class
