+++ src/main/java/bisq/desktop/main/portfolio/openoffer/OpenOffersViewModel.java
@@ -62,6 +62,10 @@ class OpenOffersViewModel extends ActivatableWithDataModel<OpenOffersDataModel> 
         dataModel.onRemoveOpenOffer(openOffer, resultHandler, errorMessageHandler);
     }
 
+    void onEditOpenOffer(OpenOffer openOffer, ResultHandler resultHandler, ErrorMessageHandler errorMessageHandler) {
+        dataModel.onEditOpenOffer(openOffer, resultHandler, errorMessageHandler);
+    }
+
     public ObservableList<OpenOfferListItem> getList() {
         return dataModel.getList();
     }
